requirements.txt
pydantic
python-dotenv
fastapi
uvicorn
ollama